+++ src/core/media/player_implementation.cpp
@@ -552,7 +552,7 @@ media::PlayerImplementation<Parent>::PlayerImplementation(const media::PlayerImp
         Parent::seeked_to()(value);
     });
 
-    d->engine->on_buffering_changed_signal().connect([this](int64_t value)
+    d->engine->on_buffering_changed_signal().connect([this](uint32_t value)
     {
         std::cout << __PRETTY_FUNCTION__ << "buffering" << value << std::endl;
         Parent::buffering_changed()(value);
